+++ package.json
@@ -47,8 +47,7 @@
     "react-native-web": "~0.19.10",
     "react-dom": "18.2.0",
     "@expo/metro-runtime": "~3.2.3",
-    "expo-dev-client": "~4.0.28",
-    "expo-random": "~14.0.1"
+    "expo-dev-client": "~4.0.28"
   },
   "devDependencies": {
     "@babel/core": "^7.24.0",
